src/tests/test_pipeline.py
#!/usr/bin/env python3
"""
Test script to verify the pipeline logic without heavy dependencies
"""
import yaml
import sys
import os

def load_config():
    # 从src/tests/目录访问根目录的configs/config.yaml
    config_path = os.path.join(os.path.dirname(__file__), '../../configs/config.yaml')
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)

def test_data_loading():
    """Test that we can load the sample data"""
    config = load_config()
    # 修正数据路径，从src/tests/目录访问
    base_path = os.path.join(os.path.dirname(__file__), '../..')
    train_path = os.path.join(base_path, config['data']['train_path'])
    
    if not os.path.exists(train_path):
        print(f"❌ Training data not found at {train_path}")
        return False
    
    with open(train_path, 'r') as f:
        lines = f.readlines()
    
    print(f"✅ Successfully loaded {len(lines)} events from {train_path}")
    
    # Check first few lines
    for i, line in enumerate(lines[:3]):
        parts = line.strip().split()
        if len(parts) != 5:
            print(f"❌ Invalid data format in line {i+1}: {line.strip()}")
            return False
        print(f"   Event {i+1}: x={parts[0]}, y={parts[1]}, t={parts[2]}, p={parts[3]}, label={parts[4]}")
    
    return True

def test_feature_extraction():
    """Test feature extraction logic"""
    print("\n--- Testing Feature Extraction ---")
    
    # Mock event data: x, y, t, p
    raw_events = [
        [100, 150, 1000, 1],
        [101, 150, 1100, 1], 
        [99, 149, 1200, 0]
    ]
    
    config = load_config()
    h, w = config['data']['resolution_h'], config['data']['resolution_w']
    output_dim = config['model']['input_feature_dim']
    
    print(f"Image resolution: {h} x {w}")
    print(f"Feature dimension: {output_dim}")
    
    # Simulate feature extraction
    feature_sequence = []
    p_time_map = [[0.0] * w for _ in range(h)]
    n_time_map = [[0.0] * w for _ in range(h)]
    
    for i, (x, y, t, p) in enumerate(raw_events):
        x_norm = x / w
        y_norm = y / h
        
        if i > 0:
            dt = t - raw_events[i-1][2]
        else:
            dt = 0
            
        # Polarity-aware time since last event at this pixel
        if p > 0:
            dt_pixel = t - p_time_map[int(y)][int(x)]
            p_time_map[int(y)][int(x)] = t
        else:
            dt_pixel = t - n_time_map[int(y)][int(x)]
            n_time_map[int(y)][int(x)] = t
        
        # Create feature vector
        features = [x_norm, y_norm, t, p, dt, dt_pixel] + [0.0] * (output_dim - 6)
        feature_sequence.append(features)
        
        print(f"   Event {i+1}: features={features[:6]}... (total {len(features)} features)")
    
    print("✅ Feature extraction logic works correctly")
    return True

def test_model_architecture():
    """Test model architecture without PyTorch"""
    print("\n--- Testing Model Architecture ---")
    
    config = load_config()
    model_config = config['model']
    
    print(f"Input feature dim: {model_config['input_feature_dim']}")
    print(f"d_model: {model_config['d_model']}")
    print(f"n_layers: {model_config['n_layers']}")
    print(f"d_state: {model_config['d_state']}")
    
    # Simulate forward pass dimensions
    batch_size = config['training']['batch_size']
    seq_len = config['data']['sequence_length']
    
    print(f"\nSimulated forward pass:")
    print(f"  Input: [{batch_size}, {seq_len}, {model_config['input_feature_dim']}]")
    print(f"  After embedding: [{batch_size}, {seq_len}, {model_config['d_model']}]")
    print(f"  After {model_config['n_layers']} Mamba layers: [{batch_size}, {seq_len}, {model_config['d_model']}]")
    print(f"  Output: [{batch_size}, {seq_len}, 1]")
    
    print("✅ Model architecture looks correct")
    return True

def main():
    print("🧪 Testing Event Flare Removal Pipeline")
    print("=" * 50)
    
    tests = [
        ("Data Loading", test_data_loading),
        ("Feature Extraction", test_feature_extraction), 
        ("Model Architecture", test_model_architecture)
    ]
    
    results = []
    for test_name, test_func in tests:
        print(f"\n--- Testing {test_name} ---")
        try:
            result = test_func()
            results.append((test_name, result))
        except Exception as e:
            print(f"❌ {test_name} failed with error: {e}")
            results.append((test_name, False))
    
    print("\n" + "=" * 50)
    print("📊 Test Results:")
    all_passed = True
    for test_name, passed in results:
        status = "✅ PASS" if passed else "❌ FAIL"
        print(f"  {test_name}: {status}")
        if not passed:
            all_passed = False
    
    if all_passed:
        print("\n🎉 All tests passed! The pipeline logic is working correctly.")
        print("📝 Next steps:")
        print("   1. Install PyTorch: pip install torch")
        print("   2. Install other dependencies: pip install numpy scikit-learn tqdm")
        print("   3. Install Mamba: pip install mamba-ssm")
        print("   4. Run: python main.py --config configs/config.yaml")
    else:
        print("\n❌ Some tests failed. Please fix the issues above.")
        return 1
    
    return 0

if __name__ == "__main__":
    sys.exit(main())