requirements.txt
torch>=2.0.0
numpy>=1.21.0
pyyaml>=6.0
tqdm>=4.64.0
scikit-learn>=1.1.0